requirements.txt
requests
dnspython
slacker